Add files via upload
Question 1
---------------
You are given an array prices where prices[i] is the price of a given stock on the ith day.
You want to maximize your profit by choosing a single day to buy one stock and choosing a different day in the future to sell that stock.
Return the maximum profit you can achieve from this transaction. If you cannot achieve any profit, return 0.

Example 1:
Input: prices = [7,1,5,3,6,4]
Output: 5
Explanation: Buy on day 2 (price = 1) and sell on day 5 (price = 6), profit = 6-1 = 5.
Note that buying on day 2 and selling on day 1 is not allowed because you must buy before you sell.

Example 2:
Input: prices = [7,6,4,3,1]
Output: 0
Explanation: In this case, no transactions are done and the max profit = 0.

Question 2
---------------
'^' means power. (2^5) = Math.pow(2,5) = 32

The series,
(1^1)+(2^2)+(3^3)+(4^4) ... (10^10) = 10405071317.

Find the last ten digits of the series.
(1^1)+(2^2)+(3^3)+ ...... +(1000^1000).